Move ignored test to the appropriate location.
