+++ server/src/test/java/org/spine3/server/entity/RepositoryShould.java
@@ -25,7 +25,6 @@ import com.google.common.base.Predicates;
 import com.google.common.collect.Lists;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.storage.RecordStorage;
@@ -255,26 +254,5 @@ public class RepositoryShould {
     private void createAndStore(String entityId) {
         ProjectEntity entity = repository.create(createId(entityId));
         repository.store(entity);
-    }
-
-    @Ignore
-    // TODO:2017-05-03:dmytro.dashenkov: This test is still valid for the non-record-based repos.
-    @Test(expected = IllegalStateException.class)
-    public void throw_ISE_if_unable_to_load_entity_by_id_from_storage_index() {
-        createAndStoreEntities();
-
-        // Store a troublesome entity, which cannot be loaded.
-        final TenantAwareOperation op = new TenantAwareOperation(tenantId) {
-            @Override
-            public void run() {
-                createAndStore(TestRepo.troublesome.getId());
-            }
-        };
-        op.execute();
-
-        final Iterator<ProjectEntity> iterator = getIterator(tenantId);
-
-        // This should iterate through all.
-        Lists.newArrayList(iterator);
     }
 }
